add support for SCO fake sink suspend prevention
This commit brings the code from Canonical which make the sink,
depending on the output selected, prevent the suspension of the
Bluetooth headset. This is forward-ported from the old version using
the code from [1].

For this to work properly, it also requires a custom patch in the
Pulseaudio's bluez module.

[1]
  - https://github.com/saemy/pulseaudio/commit/4bb6142da20b7f38c2a343a4f007a3c0c2787b97
  - https://github.com/saemy/pulseaudio/commit/60d26f12dc67af7f634aa4e5f361e7562ec050b1

[PATCH] sink: move fake SCO property setting to a hook instead

Pulseaudio calls the sink's set_port callback on IO thread instead of
main thread if "deferred volumes" are in use, and with 65483884 ("sink:
Enable deferred volume for hw volume control"), we're doing just that.
This means we cannot update SCO fake sink's property from that callback,
as an assert will cause the whole PA to crash.

Thus, move the property setting logic into a hook, which is guaranteed
to run on the main thread. Yes, I know that it's a little bit silly to
run hook for our own sink, but that's the way it has to be.

Note that we register a separated hook from the existing `sink_port_
changed_hook_cb`, since that one is used for another purpose and has
some hook/unhook logic which I would rather not touch.

Version 2:
  - Account for the fact that `data->device_port` can be null, presumbly
    when switching to a parking port.

Fixes: https://github.com/ubports/ubuntu-touch/issues/1977